Move test file to subdirectory
On the CI machine, the test_e1.py file was next to the developement
e1.py file, which meant the globally-installed e1.py file wasn't being
imported.  Since the C shared object library is located relative to the
imported e1.py file, the wrong SO location was being chosen, resulting
in a failure.  The test running in a subdirectory will miss the dev
e1.py file, and find the global e1.py module.